+++ FitnessFoodsLC/FitnessFoodsLC.Context/Migrations/FitnessFoodsLCContextModelSnapshot.cs
@@ -24,11 +24,11 @@ namespace FitnessFoodsLC.Context.Migrations
 
             modelBuilder.Entity("FitnessFoodsLC.Domain.Products.Product", b =>
                 {
-                    b.Property<int>("Id")
+                    b.Property<long>("Id")
                         .ValueGeneratedOnAdd()
-                        .HasColumnType("integer");
+                        .HasColumnType("bigint");
 
-                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<int>("Id"));
+                    NpgsqlPropertyBuilderExtensions.UseIdentityByDefaultColumn(b.Property<long>("Id"));
 
                     b.Property<string>("Barcode")
                         .HasColumnType("text");
@@ -39,8 +39,8 @@ namespace FitnessFoodsLC.Context.Migrations
                     b.Property<string>("Categories")
                         .HasColumnType("text");
 
-                    b.Property<int?>("Code")
-                        .HasColumnType("integer");
+                    b.Property<long?>("Code")
+                        .HasColumnType("bigint");
 
                     b.Property<string>("ImageUrl")
                         .HasColumnType("text");
